move map_sdss to oldviews.py ... only used during sdss coadd construction
